chore: heavy refactor
