+++ src/pages/main/Main.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Header, NameCanvas, Presentation } from '../../shared/components';
+import { Header, Canvas, Presentation } from '../../shared/components';
 
 import styles from './Main.css';
 
@@ -8,7 +8,7 @@ const Main = () => {
 	return (
 		<div className={ styles.root }>
 			<Header />
-			<NameCanvas isTouchable={ false } />
+			<Canvas isTouchable={ false } />
 			<Presentation />
 		</div>
 	);
