+++ src/shared/components/canvas/components/index.js
@@ -1,0 +1,4 @@
+export { default as Camera } from './camera/Camera';
+export { default as Lighting } from './lighting/Lighting';
+export { default as Name } from './name/Name';
+export { default as Projects } from './projects/Projects';
