+++ src/shared/components/canvas/components/name/helpers/Text.js
