--- src/shared/components/canvas/extended/index.js
@@ -1,3 +1,0 @@
-export { default as ProjectsCanvas } from './projects/Projects';
-export { default as NameCanvas } from './name/Name';
-
