+++ src/shared/components/index.js
@@ -1,7 +1,3 @@
 export { default as Canvas } from './canvas';
-export {
-	NameCanvas,
-	ProjectsCanvas,
-} from './canvas/extended';
 export { default as Header } from './header';
 export { default as Presentation } from './presentation';
