+++ src/shared/utils/index.js
@@ -1,12 +1,3 @@
-const zDepthFinder = (canvasHeight, fov ) => {
-	const splitAngle = fov/2;
-
-	// Use tangent formula to find adjacent side
-	const result = -(canvasHeight) / (Math.abs(Math.tan(splitAngle*Math.PI/180)) * 2);
-
-	return result;
-};
-
 const generateVectors = (number, zDepth, peak) => {
 	const vectors = [];
 
@@ -42,7 +33,6 @@ const gaussianFunction = (x, peakBase, stdDev) => {
 };
 
 export {
-	zDepthFinder,
 	generateVectors,
 	gaussianFunction,
 };
